Optimise stop media event position.
We don't want media to be stopped if the page
is the same, and it not stopped if the app is reset (because
it will have had to stop naturally for that to happen). We only
want to stop media if media is running and some other page
is navigated to.