Refactor video component
This changes it to a fullscreen video player. The change was
made because we always want the video fullscreen when it plays.

The video player has also been changed to one with a simpler more
standard API, and no controls (so we can add our own).

We also use a video provider to control the state of the video container
as this has to exist fullscreen in the DOM for the player to be able to run
fullscreen.