Refactor the Gloss component
This is a conversion of the Gloss component to a functional component,
and to make it show and hide based on signals via the gloss context.
It replaces the previous method of storing the class in context and calling
methods on it, which is not a very React way of doing thing and resulted
in extra unneeded renders.