Catch any drag event in the overlay
We capture drag events in the overlay to stop any leakage of these events to parent layers.

This is needed if a draggable component is nested in the overlay, but its drag
binding does not extend to the edges of the overlay. This will almost always
be true (due to other nav, etc)